+++ pyvkfft/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+# PyVkFFT
+#   (c) 2021- : ESRF-European Synchrotron Radiation Facility
+#       authors:
+#         Vincent Favre-Nicolin, favre@esrf.fr
+
+from .simple import fftn, ifftn, rfftn, irfftn, clear_vkfftapp_cache